Add paren expr.
We need to take care of the terminating condition for the loop. In this
case we modify infix_binding_power to return None when encountering an
unrecognised operator. We modify the loop to break on this condition
which in turns returns us up the call stack (to the previous expr).